finish AtlasTest, fix ActionsTest

+++ cocos2d-android/src/org/cocos2d/actions/interval/CCFadeIn.java
@@ -18,7 +18,7 @@ public class CCFadeIn extends CCIntervalAction {
 
     @Override
     public void update(float t) {
-        ((CCRGBAProtocol) target).setOpacity((byte) (255.0f * t));
+        ((CCRGBAProtocol) target).setOpacity((int) (255.0f * t));
     }
 
     @Override
